Add snappshot tests for ShuffleToggleButton and remove UI tests that had the same purpose
