--- media-ui/src/androidTest/java/com/google/android/horologist/media/ui/components/controls/ShuffleToggleButtonTest.kt
@@ -1,62 +1,0 @@
-/*
- * Copyright 2022 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:OptIn(ExperimentalHorologistMediaUiApi::class)
-
-package com.google.android.horologist.media.ui.components.controls
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Shuffle
-import androidx.compose.material.icons.filled.ShuffleOn
-import androidx.compose.ui.test.junit4.createComposeRule
-import com.google.android.horologist.media.ui.ExperimentalHorologistMediaUiApi
-import com.google.android.horologist.test.toolbox.matchers.hasIconImageVector
-import org.junit.Rule
-import org.junit.Test
-
-class ShuffleToggleButtonTest {
-
-    @get:Rule
-    val composeTestRule = createComposeRule()
-
-    @Test
-    fun givenShuffleIsOn_thenIconIsShuffleOn() {
-        // given
-        composeTestRule.setContent {
-            ShuffleToggleButton(
-                onToggle = {},
-                shuffleOn = true
-            )
-        }
-
-        // then
-        composeTestRule.onNode(hasIconImageVector(Icons.Default.ShuffleOn)).assertExists()
-    }
-
-    @Test
-    fun givenShuffleIsOff_thenIconIsShuffle() {
-        // given
-        composeTestRule.setContent {
-            ShuffleToggleButton(
-                onToggle = {},
-                shuffleOn = false
-            )
-        }
-
-        // then
-        composeTestRule.onNode(hasIconImageVector(Icons.Default.Shuffle)).assertExists()
-    }
-}
