+++ media-ui/src/test/snapshots/images/com.google.android.horologist.media.ui.controls_ShuffleToggleButtonTest_givenShuffleIsOff_thenIconIsShuffle.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ccecce5acb7b0bb7ffaf4c04202bbf4ffd156b1fa8145c835d79d0705f28e53d
+size 26706
