+++ media-ui/src/test/snapshots/images/com.google.android.horologist.media.ui.controls_ShuffleToggleButtonTest_givenShuffleIsOn_thenIconIsShuffleOn.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f7a83a37ba8d8d1648d27ca7ccc6b214a540d25000177529e6879527126417da
+size 28430
